+++ django_nested_values/__init__.py
@@ -1,11 +1,11 @@
 """Django Nested Values - Enable .prefetch_related().values_nested() in Django ORM."""
 
 from django_nested_values.queryset import (
-    NestedObject,
+    AttrDict,
     NestedValuesQuerySet,
     NestedValuesQuerySetMixin,
     RelatedList,
 )
 
-__all__ = ["NestedObject", "NestedValuesQuerySet", "NestedValuesQuerySetMixin", "RelatedList"]
+__all__ = ["AttrDict", "NestedValuesQuerySet", "NestedValuesQuerySetMixin", "RelatedList"]
 __version__ = "0.5.2"